requirements.txt
# ── Frontend ──────────────────────────
streamlit>=1.32.0
requests==2.31.0

# ── Backend API ───────────────────────
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic>=2.6
# ── AI / LLM ──────────────────────────
langchain==0.3.25
langchain-core==0.3.83
langchain-community==0.3.24
langchain-google-genai==2.0.7
chromadb==0.5.0
sentence-transformers==2.7.0

# ── Document Parsing ──────────────────
pymupdf==1.24.3
python-docx==1.1.2

# ── Config ────────────────────────────
python-dotenv==1.0.1